refactor!: move htsget structs to config, and resolve storage type in config
