+++ deploy/bin/htsget-lambda.ts
@@ -1,7 +1,7 @@
 #!/usr/bin/env node
 import "source-map-support/register";
 import * as cdk from "aws-cdk-lib";
-import { HtsgetLambdaStack } from "../lib/htsget-lambda-stack";
+import {HtsgetLambdaStack} from "../lib/htsget-lambda-stack";
 
 export const STACK_NAME = "HtsgetLambdaStack";
 const STACK_DESCRIPTION = "A stack deploying htsget-lambda with API gateway.";
