+++ htsget-actix/benches/request_benchmarks.rs
@@ -10,8 +10,8 @@ use reqwest::blocking::ClientBuilder;
 use reqwest::Result;
 use serde::{Deserialize, Serialize};
 
+use htsget_config::types::{Headers, JsonResponse};
 use htsget_http::{PostRequest, Region};
-use htsget_search::htsget::{Headers, JsonResponse};
 use htsget_test::http_tests::{default_config_fixed_port, default_dir, default_dir_data};
 
 const REFSERVER_DOCKER_IMAGE: &str = "ga4gh/htsget-refserver:1.5.0";
