+++ htsget-config/Cargo.toml
@@ -15,6 +15,8 @@ s3-storage = []
 default = ["s3-storage"]
 
 [dependencies]
+thiserror = "1.0"
+async-trait = "0.1"
 noodles = { version = "0.32", features = ["core"] }
 serde = { version = "1.0", features = ["derive"] }
 serde_with = "2.2"
