+++ htsget-http/src/error.rs
@@ -2,7 +2,7 @@ use http::StatusCode;
 use serde::Serialize;
 use thiserror::Error;
 
-use htsget_search::htsget::HtsGetError as HtsGetSearchError;
+use htsget_config::types::HtsGetError as HtsGetSearchError;
 
 pub type Result<T> = core::result::Result<T, HtsGetError>;
