+++ htsget-http/src/post_request.rs
@@ -1,7 +1,7 @@
 use serde::{Deserialize, Serialize};
 use tracing::instrument;
 
-use htsget_config::Query;
+use htsget_config::types::Query;
 
 use crate::{QueryBuilder, Result};
 
@@ -61,7 +61,7 @@ impl PostRequest {
 
 #[cfg(test)]
 mod tests {
-  use htsget_config::{Class, Format};
+  use htsget_config::types::{Class, Format};
 
   use super::*;
