+++ htsget-lambda/src/handlers/mod.rs
@@ -6,8 +6,8 @@ use lambda_http::http::{header, StatusCode};
 use serde::Serialize;
 use serde_json::Error;
 
+use htsget_config::types::JsonResponse;
 use htsget_http::{HtsGetError, Result};
-use htsget_search::htsget::JsonResponse;
 
 use crate::{Body, Response};
