+++ htsget-search/benches/search_benchmarks.rs
@@ -5,12 +5,11 @@ use criterion::{criterion_group, criterion_main, BenchmarkGroup, Criterion};
 use tokio::runtime::Runtime;
 
 use htsget_config::config::cors::CorsConfig;
-use htsget_config::Class::Header;
-use htsget_config::Format::{Bam, Bcf, Cram, Vcf};
-use htsget_config::Query;
+use htsget_config::types::Class::Header;
+use htsget_config::types::Format::{Bam, Bcf, Cram, Vcf};
+use htsget_config::types::{HtsGetError, Query};
 use htsget_search::htsget::from_storage::HtsGetFromStorage;
 use htsget_search::htsget::HtsGet;
-use htsget_search::htsget::HtsGetError;
 use htsget_search::storage::data_server::HttpTicketFormatter;
 
 const BENCHMARK_DURATION_SECONDS: u64 = 30;
