+++ htsget-search/src/htsget/bcf_search.rs
@@ -16,10 +16,8 @@ use tokio::io::AsyncRead;
 use tracing::{instrument, trace};
 
 use crate::htsget::search::{find_first, BgzfSearch, BinningIndexExt, Search};
-use crate::{
-  htsget::{Format, Query, Result},
-  storage::{BytesPosition, Storage},
-};
+use crate::storage::{BytesPosition, Storage};
+use crate::{Format, Query, Result};
 
 type AsyncReader<ReaderType> = bcf::AsyncReader<bgzf::AsyncReader<ReaderType>>;
 
@@ -135,9 +133,9 @@ mod tests {
   #[cfg(feature = "s3-storage")]
   use crate::htsget::from_storage::tests::with_aws_storage_fn;
   use crate::htsget::from_storage::tests::with_local_storage_fn;
-  use crate::htsget::{Class::Header, Headers, HtsGetError::NotFound, Response, Url};
   use crate::storage::data_server::HttpTicketFormatter;
   use crate::storage::local::LocalStorage;
+  use crate::{Class::Header, Headers, HtsGetError::NotFound, Response, Url};
 
   use super::*;
