+++ htsget-search/src/lib.rs
@@ -1,6 +1,9 @@
 pub use htsget_config::config::{Config, DataServerConfig, ServiceInfo, TicketServerConfig};
-pub use htsget_config::resolver::{QueryAllowed, RegexResolver, Resolver};
+pub use htsget_config::resolver::{IdResolver, QueryAllowed, Resolver};
 pub use htsget_config::storage::Storage;
+pub use htsget_config::types::{
+  Class, Format, Headers, HtsGetError, JsonResponse, Query, Response, Result, Url,
+};
 
 pub mod htsget;
 pub mod storage;
