+++ htsget-search/src/storage/mod.rs
@@ -10,7 +10,6 @@ use std::time::Duration;
 use async_trait::async_trait;
 use base64::engine::general_purpose;
 use base64::Engine;
-use http::uri::Authority;
 use http::{uri, HeaderValue};
 use thiserror::Error;
 use tokio::io::AsyncRead;
@@ -18,9 +17,10 @@ use tower_http::cors::{AllowHeaders, AllowMethods, AllowOrigin, CorsLayer, Expos
 use tracing::instrument;
 
 use htsget_config::config::cors::CorsConfig;
-use htsget_config::{Class, Scheme};
+use htsget_config::storage::local::LocalStorage;
+use htsget_config::types::{Class, Scheme};
 
-use crate::htsget::{Headers, Url};
+use crate::{Headers, Url};
 
 #[cfg(feature = "s3-storage")]
 pub mod aws;
@@ -104,15 +104,15 @@ pub enum StorageError {
   AwsS3Error(String, String),
 }
 
-impl UrlFormatter for (Scheme, Authority, String) {
+impl UrlFormatter for LocalStorage {
   fn format_url<K: AsRef<str>>(&self, key: K) -> Result<String> {
     uri::Builder::new()
-      .scheme(match self.0 {
+      .scheme(match self.scheme() {
         Scheme::Http => uri::Scheme::HTTP,
         Scheme::Https => uri::Scheme::HTTPS,
       })
-      .authority(self.1.to_string())
-      .path_and_query(format!("{}/{}", self.2, key.as_ref()))
+      .authority(self.authority().to_string())
+      .path_and_query(format!("{}/{}", self.path_prefix(), key.as_ref()))
       .build()
       .map_err(|err| StorageError::InvalidUri(err.to_string()))
       .map(|value| value.to_string())
