+++ htsget-test/src/server_tests.rs
@@ -1,12 +1,13 @@
-use base64::engine::general_purpose;
-use base64::Engine;
 use std::fmt::Debug;
 use std::net::SocketAddr;
 use std::str::FromStr;
 
+use crate::util::expected_bgzf_eof_data_url;
+use base64::engine::general_purpose;
+use base64::Engine;
 use futures::future::join_all;
 use futures::TryStreamExt;
-use htsget_config::{Class, Format};
+use htsget_config::types::Format;
 use http::header::HeaderName;
 use http::{HeaderMap, HeaderValue, Method};
 use noodles_bgzf as bgzf;
@@ -15,8 +16,9 @@ use reqwest::ClientBuilder;
 use serde::Deserialize;
 use serde_json::{json, Value};
 
+use htsget_config::types::Class;
+
 use crate::http_tests::{Header, Response, TestRequest, TestServer};
-use crate::util::expected_bgzf_eof_data_url;
 use crate::Config;
 
 /// Test response with with class.
